03-openai-api-summarizer/requirements.txt
openai
python-dotenv
PyPDF2